+++ CHANGELOG.md
@@ -9,6 +9,7 @@ meaning as `!@wlambda`.
 to prohibit it to recurse into certain keys.
 * Feature: Added positive value condition to recursive selectors `$S(**=:{x=10})`
 to select only specific values to recurse into.
+* Feature: Added std:bytes:replace function.
 
 0.6.1 (2020-06-04)
 ==================